GH-2483 Validating that README exists before deleting it

+++ scripts/publish-builds.sh
@@ -55,7 +55,9 @@ publish_snapshot() {
 
     cp -R $LIB_DIR/* $TMP_DIR/
 
-    rm README.md
+    if [ -e README.md ]; then
+       rm README.md
+    fi
 
     echo `date` > $TMP_DIR/BUILD_INFO
     echo $SHA >> $TMP_DIR/BUILD_INFO
